+++ packages/node/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @module-federation/node
 
+## 2.7.11
+
+### Patch Changes
+
+- Updated dependencies [609d477]
+- Updated dependencies [0ab51b8]
+- Updated dependencies [98a29c3]
+- Updated dependencies [f6381e6]
+  - @module-federation/runtime@0.18.0
+  - @module-federation/enhanced@0.18.0
+  - @module-federation/sdk@0.18.0
+
 ## 2.7.10
 
 ### Patch Changes